+++ contracts/mock/libs/data-structures/IncrementalMerkleTreeMock.sol
@@ -67,6 +67,30 @@ contract IncrementalMerkleTreeMock {
         return _addressTree.root();
     }
 
+    function verifyUintProof(
+        bytes32[] memory siblings_,
+        uint256 directionBits_,
+        bytes32 element_
+    ) external view returns (bool) {
+        return _uintTree.verifyProof(siblings_, directionBits_, element_);
+    }
+
+    function verifyBytes32Proof(
+        bytes32[] memory siblings_,
+        uint256 directionBits_,
+        bytes32 element_
+    ) external view returns (bool) {
+        return _bytes32Tree.verifyProof(siblings_, directionBits_, element_);
+    }
+
+    function verifyAddressProof(
+        bytes32[] memory siblings_,
+        uint256 directionBits_,
+        bytes32 element_
+    ) external view returns (bool) {
+        return _addressTree.verifyProof(siblings_, directionBits_, element_);
+    }
+
     function getUintTreeHeight() external view returns (uint256) {
         return _uintTree.height();
     }